+++ game/controller/roleController.go
@@ -28,7 +28,7 @@ func innerPlayerLogin(roleId int64, request proto.Message) {
 		Sid:    loginRequest.Sid,
 		RoleId: loginRequest.RoleId,
 	}
-	client.GetInnerClient(client.WORLD).SendInnerMsg(int32(protoGen.InnerProtoCode_INNER_LOGIN_REQ), loginRequest.RoleId, innerLoginReq)
+	client.GetInnerClient(client.WORLD).SendInnerMsg(protoGen.InnerProtoCode_INNER_LOGIN_REQ, loginRequest.RoleId, innerLoginReq)
 	gameRole := role.NewRole(loginRequest.RoleId, nil)
 	RoleOlineMgr.AddRole(gameRole)
 }
@@ -41,7 +41,7 @@ func loginResponseFromWorldServer(roleId int64, request proto.Message) {
 		log.Infof(" role id = %d not found or have disconnected", roleId)
 		return
 	}
-	client.GetInnerClient(client.GATE_WAY).SendInnerMsg(int32(protoGen.InnerProtoCode_INNER_LOGIN_RES), roleId, innerLoginResponse)
+	client.GetInnerClient(client.GATE_WAY).SendInnerMsg(protoGen.InnerProtoCode_INNER_LOGIN_RES, roleId, innerLoginResponse)
 }
 
 func innerPlayerDisconnect(roleId int64, request proto.Message) {
@@ -56,5 +56,5 @@ func innerPlayerDisconnect(roleId int64, request proto.Message) {
 		return
 	}
 	log.Infof("roleId =%d logout", roleId)
-	client.GetInnerClient(client.GAME).SendInnerMsg(int32(protoGen.InnerProtoCode_INNER_PLAYER_DISCONNECT_REQ), roleId, playerDisconnectRequest)
+	client.GetInnerClient(client.GAME).SendInnerMsg(protoGen.InnerProtoCode_INNER_PLAYER_DISCONNECT_REQ, roleId, playerDisconnectRequest)
 }
